requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
decorator==4.2.1
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.2
Pillow==5.1.0
psutil==5.4.5
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
subprocess32==3.2.7
